A proposed fix to #58.  In ServerBase.RequestQueue there is a risk of deadlock in ScheduleIncomingRequest where whilst inside the lock we call OperationCompleted. This change moves that OperationCompleted call to outside the lock.
The problem is that whilst inside the RequestQueue lock we called OperationCompleted that sent data via the TcpChannel, this tries to get the channels 'DataLock'.
If other requests come in they take the TcpChannel lock first, then end up waiting on the RequestQueue lock.  We should probably avoid calling outside of the RequestQueue class
whilst holding the lock.